+++ CHANGELOG.md
@@ -6,9 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## Unreleased
 ### Added
 - (Meta,Core,Html,Drupal,Twig,Site,Ui) Started changelog, mirrored out to components on release.
+- (Twig,Drupal) Parse Twig comments for @Component metadata comments.
 
 ### Fixed
 - (Demo) #65, $66 Fixed missing semicolon and incorrect version constraints.
+
+### Deprecated
+- (Twig,Drupal) Twig component metadata using the componentinfo block has been deprecated due to issues with guarded blocks in child templates.  Please use the new comment syntax instead.
 
 ## [1.0.0] - 2017-10-04
 ### Added